+++ R/readSurvey.R
@@ -1,5 +1,5 @@
 #   Download qualtrics data into R
-#    Copyright (C) 2017 Jasper Ginn
+#    Copyright (C) 2018 Jasper Ginn
 
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -23,7 +23,6 @@
 #'Variable labels are stored as attributes.
 #'
 #' @param file_name String. A csv data file.
-#' @param convertStandardColumns Logical. If TRUE, then the function will convert general data columns (first name, last name, lat, lon, ip address, startdate, enddate etc.) to their proper format. defaults to TRUE.
 #' @param stripHTML Logical. If TRUE, then remove html tags. Defaults to TRUE
 #' @param legacyFormat Logical. If TRUE, then import "legacy" format csv files (as of 2017). Defaults to FALSE
 #'
@@ -31,6 +30,8 @@
 #' @importFrom utils read.csv
 #' @importFrom sjlabelled set_label
 #' @importFrom jsonlite fromJSON
+#' @importFrom stringr str_match
+#' @importFrom readr read_csv
 #' @return A data frame. Variable labels are stored as attributes. They are not printed on
 #' the console but are visibile in the RStudio viewer.
 #' @export
@@ -40,7 +41,6 @@
 #' }
 
 readSurvey <- function(file_name,
-                       convertStandardColumns = TRUE,
                        stripHTML = TRUE,
                        legacyFormat = FALSE) {
   # check if file exists
@@ -48,15 +48,19 @@ readSurvey <- function(file_name,
   # skip 2 rows if legacyFormat, else 3 when loading the data
   skipNr <- ifelse(legacyFormat, 2, 3)
   # import data including variable names (row 1) and variable labels (row 2)
-  rawdata <- readr::read_csv(file = file_name,
+  rawdata <- suppressMessages(readr::read_csv(file = file_name,
                              col_names = FALSE,
-                             skip = skipNr)
-  header <- readr::read_csv(file = file_name,
+                             skip = skipNr))
+  # Need contingency when 0 rows
+  assertthat::assert_that(nrow(rawdata) > 0,
+                          msg="The survey you are trying to import has no responses.")
+  # Load headers
+  header <- suppressMessages(readr::read_csv(file = file_name,
                             col_names = TRUE,
-                            n_max = 1)
+                            n_max = 1))
   # make them data.frame's, else the factor conversion in `inferDataTypes` crashes
-  rawdata <- as.data.frame(rawdata)
-  header <- as.data.frame(header)
+  #rawdata <- as.data.frame(rawdata)
+  #header <- as.data.frame(header)
   # Add names
   names(rawdata) <- names(header)
 
@@ -88,9 +92,5 @@ readSurvey <- function(file_name,
   # add variable labels
   # -------------------
   rawdata <- sjlabelled::set_label(rawdata, unlist(subquestions))
-  # Add types
-  if(convertStandardColumns) {
-    rawdata <- inferDataTypes(rawdata)
-  }
   return(rawdata)
 }