+++ R/registerOptions.R
@@ -1,5 +1,5 @@
 #   Download qualtrics data into R
-#    Copyright (C) 2017 Jasper Ginn
+#    Copyright (C) 2018 Jasper Ginn
 
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -30,6 +30,8 @@
 #'
 #' @author Jasper Ginn
 #' @importFrom yaml yaml.load_file
+#' @importFrom assertthat assert_that
+#' @importFrom assertthat is.flag
 #' @export
 #' @examples
 #' \dontrun{